+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.0.0-rc.2
-##  XX/XX/2017
+##  01/16/2017
 
 1. [](#new)
     * Allow to change the path for the `git` binary (#1)